requirements.txt
# Backend core
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20

# Database (SQLite — built-in, no extra driver needed)
SQLAlchemy==2.0.41

# AI / ML
groq>=0.12.0
scikit-learn==1.6.1
joblib==1.5.1
numpy==2.3.0
scipy==1.15.3

# Auth
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==5.0.0

# Data validation
pydantic==2.11.4
pydantic[email]==2.11.4
annotated-types==0.7.0

# PDF generation
reportlab==4.4.1
pillow==11.2.1

# Utilities
python-dotenv==1.1.0
charset-normalizer==3.4.2
